+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 * Updated some styling
 * Split the `allSites()` section setting into `allSitesUnique` and `allSitesDistinct`
 * Deprecated the `allSites()` section setting
-* Deprecated the `fallbackImage` plugin setting
+* Removed the `fallbackImage` plugin setting
 * The values for the `size()` section settings are now by default `tiny`, `small`, `medium`, `large` and `huge`, the
   corresponding CSS width values are now significantly smaller and take the container width into account.